Be more liberal with extra properties in schema validation
