+++ src/components/review-panel.tsx
@@ -1,20 +1,10 @@
 import { Button } from "@mui/material";
 import { useAppSelector } from "../state/hooks";
-import {slugify} from '../state/uiSpec-reducer';
+import { downloadNotebook } from "../state/localStorage";
 
 export const ReviewPanel = () => {
 
-    const state = useAppSelector((state) => state);
-
-    const downloadNotebook = () => {
-        const element = document.createElement("a");
-        const file = new Blob([JSON.stringify(state, null, 2)], { type: 'application/json' });
-        element.href = URL.createObjectURL(file);
-        const name = slugify(state.notebook.metadata.name as string);
-        element.download = `${name}.json`;
-        document.body.appendChild(element);
-        element.click();
-    };
+    const notebook = useAppSelector(state => state.notebook);
 
     return (
         <div>
@@ -22,7 +12,7 @@ export const ReviewPanel = () => {
                 then be uploaded to the Fieldmark Conductor server
                 by a admin user.</p>
 
-            <Button variant="contained" onClick={downloadNotebook}>
+            <Button variant="contained" onClick={() => downloadNotebook(notebook)}>
                 Download Notebook
             </Button>
         </div>
